🎉 Refactoring & functionality for deleting members
