+++ src/App.tsx
@@ -1,26 +1,15 @@
 import * as React from 'react'
-import {
-  getProjects,
-  getProjectsById,
-  getProjectIds,
-  createProject,
-  patchProject,
-  patchProjectApiKey,
-  deleteProject
-} from './lib/projects'
+import { getProjects, getProjectsById, getProjectIds } from './lib/projects'
 import { UserContext, initialUser } from './context/userContext'
 import { TeamsContext } from './context/teamsContext'
 import { ProjectsContext } from './context/projectsContext'
 import { loggedIn, logout } from './lib/auth'
 import { IUser, IProject, ITeam } from './types'
-import {
-  getTeamIds,
-  getTeamsById,
-  patchTeam,
-  patchTeamMember
-} from './lib/teams'
+import { getTeamIds, getTeamsById } from './lib/teams'
 import { getUser } from './lib/user'
 import Routes from './components/functional/routes'
+import teamStore from './store/team'
+import projectStore from './store/project'
 
 const initialLoadingState = {
   user: false,
@@ -77,120 +66,11 @@ const App: React.FC = () => {
   // when depedency values are changing.
   const userProviderValue = React.useMemo(() => user, [user])
   const teamsProviderValue = React.useMemo(
-    () => ({
-      teams: teamIds,
-      teamsById,
-      patchTeam: async (teamId: string, values: { [key: string]: any }) => {
-        // Patch project
-        const patchedTeam = await patchTeam(teamId, values)
-
-        // Find previous project
-        const oldTeam = teams.find((t) => t?._id === teamId)
-
-        if (oldTeam) {
-          // Set new project
-          const newTeams = [...teams]
-          newTeams[teams.indexOf(oldTeam)] = patchedTeam
-          setTeams(newTeams)
-
-          // Return project
-          return patchedTeam
-        }
-      },
-      patchTeamMember: async (
-        teamId: string,
-        userId: string,
-        values: { [key: string]: any }
-      ) => {
-        // Patch project
-        const patchedTeam = await patchTeamMember(teamId, userId, values)
-
-        // Find previous project
-        const oldTeam = teams.find((t) => t?._id === teamId)
-
-        if (oldTeam) {
-          // Set new project
-          const newTeams = [...teams]
-          newTeams[teams.indexOf(oldTeam)] = patchedTeam
-          setTeams(newTeams)
-
-          // Return project
-          return patchedTeam
-        }
-      }
-    }),
+    () => teamStore(teamIds, teamsById, teams, setTeams),
     [teamIds, teamsById, teams]
   )
   const projectsProviderValue = React.useMemo(
-    () => ({
-      projectsById,
-      projects: projectIds,
-      createProject: async (owner: string, name: string) => {
-        // Create project
-        const project = await createProject(owner, name)
-
-        // Set projects
-        setProjects([...projects, project])
-
-        // Return project
-        return project
-      },
-      patchProject: async (
-        projectId: string,
-        values: { [key: string]: any }
-      ) => {
-        // Patch project
-        const patchedProject = await patchProject(projectId, values)
-
-        // Find previous project
-        const oldProject = projects.find((p) => p?._id === projectId)
-
-        if (oldProject) {
-          // Set new project
-          const newProjects = [...projects]
-          newProjects[projects.indexOf(oldProject)] = patchedProject
-          setProjects(newProjects)
-
-          // Return project
-          return patchedProject
-        }
-      },
-      patchProjectApiKey: async (projectId: string) => {
-        // Patch project
-        const patchedProject = await patchProjectApiKey(projectId)
-
-        // Find previous project
-        const oldProject = projects.find((p) => p?._id === projectId)
-
-        if (oldProject) {
-          // Set new project
-          const newProjects = [...projects]
-          newProjects[projects.indexOf(oldProject)] = patchedProject
-          setProjects(newProjects)
-
-          // Return project
-          return patchedProject
-        }
-      },
-      deleteProject: async (projectId: string) => {
-        // Patch project
-        const isProjectDeleted = await deleteProject(projectId)
-
-        // Find previous project
-        const oldProject = projects.find((p) => p?._id === projectId)
-
-        // Set new projects
-        if (isProjectDeleted && oldProject) {
-          const newProjects = [...projects]
-          newProjects.splice(projects.indexOf(oldProject), 1)
-          setProjects(newProjects)
-
-          return true
-        }
-
-        return false
-      }
-    }),
+    () => projectStore(projectIds, projectsById, projects, setProjects),
     [projectIds, projectsById, projects]
   )
